Add movies API
Add an express route to load data about movies, using data from rotten
tomatoes.  The data is trimmed a bit to pull out the things we’d like
to see and returned in the API call.

A generic pattern to handle routes was added to the express file, and
all APIs will be served under the `/api/` URL.  A 404 is returned if no
route responds to a `/api/` URL not handled.