Save movie metadata
Most of the changes in this commit involve writing the movie metadata
that we download to our local database.  If the movie metadata exists
in the database, then we update the data rather than creating (upsert).

Since the library functions were used both within the `api` package and
outside of it, move the `lib` folder to the root `server` directory.

Include the `slug` module and our own special `slugify` version to
create unique IDs for the move metadata.