Filter movies, better mongo support
Implement a filter function on the movies we retrieve from the movies
API query.  This filtering is configurable via environment variables,
but defaults are supplied and documented in the readme.  The filter is
applied prior to saving the data to the database (or updating it).

Also improve our mongo/mongoose support by verifying the slug is max 60
characters, and that we don’t update with null/undefined values in our
object.