Add test framework, some tests
Add the Jasmine test framework, which because we’re using Babel,
requires additional setup.  The net effect is that the tests take a bit
longer to startup, but run at normal speed once the compilation is
complete.

Add a couple tests to the existing libraries: logger and slugify.  This
was mainly done to prove out the test framework.