Include stats on "save" operation
Refactor the logic within the save functions to include statistics on
how many movies were processed, and of those, how many were adds vs
updates.  This required a check on the `saveMovie` function to see if
the movie metadata exists prior to performing the update.  In the
`saveMovies` function, we now wrap the call to `saveMovie` and update
our stats each time we return.

Also move the logic of determining the ID of a movie into the `Movie`
mongoose model so we don’t leak that functionality outside of where its
necessary.  Create a test for the model as well.