List movies UI, saved and dismissed attributes
Build out `GET` call to list the movies for the UI.  We need to
translate the Movie model to our more presentable, movie UI object.

Add in `saved` and `dismissed` user interaction flags to the Movie
model.  These (hopefully) will be used to signal user interactions with
the movie metadata.