Add UI components to display movies
Add components to display the movies, along with some basic redux logic
to load the movies from the server.  Each component is wrapped with a
container which is responsible for loading the correct actions and
state.  We’ll load these movies on the index route.

Add some tests for the client side code as well, which require some
additional modules to mock some things.  We’re able to reuse mocha for
the client side tests, but still separate the execution out in the npm
scripts.