Display movie card in UI
A first pass at the movie card to display the details of the movie in
the UI.  This includes adding many of the movie attributes to the movie
component.  In building this out, some additional attributes were
needed that required changes to the movie model and download logic.

We also have to be careful and not return all movies when requested,
but set limits (along with a skip value) to the movies API.  The UI is
not yet taking advantage of this, but simply using the defaults.