Add passport, login, secure APIs
Add `passport` as a dependency, along with `passport-local` as an
authentication strategy.  Use `bcrypt` to store a hash of the password
in the database.  Create the login and logout routes, and store the
session cookie using express’ `cookie-session`.

Move the movies APIs to a secure route, and bounce requests to those
resources that have not yet authenticated.  We have a notion of a user
within the code base, but no current way to create a user through the
APIs or web app.  Instead use a simple script to create the user (since
this will probably not be something done often enough to warrant an
API).