Add login component, actions/reducer
Add the login component and connect it to the session API.  Add the
actions and reducer redux components.  We’re also sharing some CSS, so
use variables.

In some testing it was found that we had specified the cookie `maxAge`
incorrectly, so fix that in the express configuration.  Also, the
`fetch` call must include `credentials` to make sure we include cookies
in the requests/responses.