Rename app elements
Rename all the things!

To better describe what the file represents (so that each file isn’t
named `index.js`), rename most app files to include their role
(component, container, reducer, etc).  This pattern has helped in
naming things in the past, and is useful when viewing or opening a file
in an editor.

This is just rename, but will show as a ton of new files.