Load more movies
Provide a way for the movies component to load more movies.  Change the
movies component to be a dumb component, and move that logic to the
container.  The movies component now requests a function that is used
to load “more” movies.

To accomplish this, the movie actions now requests the current state,
and bases the load movies API call off of the known skip and limit
within the state.  The movies reducer then updates the skip adding to
it the limit.  (This will probably need to be changed in the future,
but works for now.)

Some CSS refactoring was done to share some code between components.
Added a `base.scss` to include generic stylings used by the components.