Dismiss movie API
Provide an API which grants the user the ability to update a movie,
specifically dismissing a movie.  This simply marks a flag on the movie
document.  Subsequent loads of movies will (by default) not include
dismissed movies.

In order to assure that we have an even playing field, all movies
should default `false` as the dismissed and saved flags (if they are
not set to begin with).