Add saved and undo dismissed movie states API
Implement the server side logic to grant the user the ability to update
a movie between the states:

* default queue
* saved
* dismissed

The user is able to do this by modifying the `saved` and `dismissed`
flags on a movie.  The saved state would take it out of the queue
state.  The dismissed state would take it out of all other states and
ready for deletion.