Add movies containers for movie states
Separate out the movies container into three separate movies containers
for each of the states:

* queue
* saved
* dismissed

Movies can flow between each state and are then shown in that
container.  Add a header to browse between the states.  We had to
change the initial load performed in the `componentWillMount` to call
specifically an initial load action rather than the load action.  This
was done because each view change would load more movies for that state
container.