Move movie action logic to reducer
Simplify the movie actions for save, dismiss, and un-dismiss by having
it dispatch the updated movie to the reducer.  The reducer can then act
on that action type given the updated movie.  This also will give us
the updated movie details to add to the correct movies list.

Since we’re relying on getting back and movieUI object, update the
movies lib to return the UI version of the object rather than the
database version.