Test application
Enable the use of a test application which can be used to test out
components contained within the main application.  We key off the
`TEST_APP` environment variable, and if set, set a different entry
point in the webpack configuration.  This spiders out and loads a
different set of files for the client bundle.  The inverse is then true
— if we do _not_ set the `TEST_APP` environment variable, we don’t load
the test application assets into the main bundle.

Nothing much in the test app as of yet, just a simple home page with a
couple of links that don’t go anywhere.  Really just a proof of concept
at this point.