Support synopsis within Movie
Add support to display the movie synopsis within the Movie component.
This required converting the Movie component into a stateful component
because the synopsis display toggles to display the full text.  So the
majority of the changes in the Movie component is whitespace changes.

Some additional styling changes required a new swatch.