Bug fixes
* Use the `binary` HTTP request encoding to correctly load the data
from the movies site.  Provide this as a configuration option.
* Reduce the skip amount in the movies state by 1 for the appropriate
queue when a movie is saved/dismissed/undismissed.  This was done
because we were altering the list in saving/dismissing/undismissing in
removing something from a queue, and would then miss some movies when
we’d request additional movies later.  By reducing the skip amount, we
don’t miss those movies.
* Fix a typo in the setup to log errors correctly.