Movie trailer API
Add an API which can “create” a movie trailer for a movie.

Now, when we load the movie details, we’ll also grab the remote ID for
the movie document.  We store this to be used later.  We also allow
storing a `trailerUrl` in the movie document.

If the user wishes to POST a request for a movie trailer, we’ll use
that remote ID to query the movie trailer API to pull down a movie
trailer URL.  We can then take that URL and update the movie document
to include this `trailerUrl`.