Movies component changes, styling updates
**Pass containers to components**

To avoid coupling with the movies component (dumb component) and the
fact that it had to render a movie container (smart component), have
the parent container (movies container) pass down the movie container
to the movies component.  This allows us to continue to have a dumb
component just render what it’s asked.

We can continue to use this pattern whenever we have a component that
must render a child container.  The parent container can pass down to
that component the selected containers which should be used in the
rendering.

**Add movies test page**

With this container / component separation, we can add a movies
component test page and provide a mock container which it will use when
rendering.

**Loading and buttons style components**

Rename `base.scss` to `buttons.scss` to better describe what was
included and to also break up these style components into separate
files.  Add a loading styling component which is used in the movies
component when we’re waiting.