"has more movies" flag
Add a boolean flag to each movies queue to indicate if more movies
exist within that section (queue, saved, dismissed).  We determine if
this is the case by checking if the amount sent to the reducer was
equal to the amount in the limit.  Doing this means that we have to
send only the new movies to the reducer, not the full list (which moves
more logic from the actions to the reducer).

With the flag in place, we can hide the loading and buttons if there
are no more movies available in the component.