Remove stale movies
Add the ability to remove movies which are both `dismissed` and have
not been modified recently (as determined by the configuration).  This
allows us to cleanup old movies which are dismissed by the user and not
updated by the movies API.

To help in this new feature, we must update the `created-modified`
mongoose plugin to get a pre update hook in place to correctly update
the movies `modified` flag.