Refactor movie reducer
Instead of using a single reducer to handle the state of all 3 movie
groups (movies queue, saved movies, dismissed movies), split the
reducer into a parent and child reducer.  The parent manages the
overall (high level) state of the movie groups, while the child manages
each group’s details.  This allows us to reduce the overall duplication
of code, and hopefully make it a bit easier to understand.

The commit diff is a bit difficult to understand, since a bulk of the
`movies.reducer` was changed, and the tests were almost fully
re-written.  Looking at the end result may be a bit easier to
understand.

Still need to refactor the actions and containers to make use of this
new system…