Refactor movie actions
There was a lot of duplicate code across the movie actions, so combine
them all into the `movie.actions.js` library.  This required the tests
to be merged as well, and there’s still a lot of those…  Remove the
unnecessary files after the code refactor.

Also wire up the new actions in the containers and test it out.  There
was one bug in the way the reducer handled a state change it didn’t
recognize, and instead of creating the initial state in that action,
create the initial state on load of the reducer.  This prohibits
clearing the state on unrecognized actions.